+++ internal/dnssvc/dnssvc.go
@@ -32,8 +32,8 @@ type DNSService struct {
 	// clientGetter is used to get the client's address from the request's
 	// context.  It's only used for testing.
 	//
-	// TODO(e.burkov):  Put custom client's address to proxy context, when it
-	// start supporting the [context.Context].  Then get rid of this interface.
+	// TODO(e.burkov):  Use custom client's address from dnsproxy context and
+	// get rid of this interface.
 	clientGetter ClientGetter
 
 	// bootstrapUpstreams is a list of upstreams to close on shutdown.
@@ -59,7 +59,7 @@ func New(conf *Config) (svc *DNSService, err error) {
 		clientGetter: conf.ClientGetter,
 		clients:      newClientStorage(clients),
 	}
-	prxConf.RequestHandler = svc
+	prxConf.RequestHandler = svc.Wrap(svc)
 
 	prx, err := proxy.New(prxConf)
 	if err != nil {
@@ -183,16 +183,33 @@ func (svc *DNSService) closeBootstraps() (errs []error) {
 }
 
 // type check
+var _ proxy.Middleware = (*DNSService)(nil)
+
+// Wrap implements the [proxy.Middleware] interface for *DNSService.
+func (svc *DNSService) Wrap(h proxy.Handler) (wrapped proxy.Handler) {
+	f := func(ctx context.Context, p *proxy.Proxy, dctx *proxy.DNSContext) (err error) {
+		// This is used to substitute the client's address in tests.
+		dctx.Addr = svc.clientGetter.Address(dctx)
+
+		// Check the address privateness because proxy does it before
+		// the substitution.  See TODO on [DNSService.clientGetter].
+		dctx.IsPrivateClient = svc.proxy.PrivateSubnets.Contains(dctx.Addr.Addr())
+
+		return h.ServeDNS(ctx, p, dctx)
+	}
+
+	return proxy.HandlerFunc(f)
+}
+
+// type check
 var _ proxy.Handler = (*DNSService)(nil)
 
 // ServeDNS implements the [proxy.Handler] interface for *DNSService.
-func (svc *DNSService) ServeDNS(ctx context.Context, p *proxy.Proxy, dctx *proxy.DNSContext) (err error) {
-	// This is used to substitute the client's address in tests.
-	dctx.Addr = svc.clientGetter.Address(dctx)
-
-	// Check the address privateness because proxy does it before substitution.
-	// See TODO on [DNSService.clientGetter].
-	dctx.IsPrivateClient = p.PrivateSubnets.Contains(dctx.Addr.Addr())
+func (svc *DNSService) ServeDNS(
+	ctx context.Context,
+	p *proxy.Proxy,
+	dctx *proxy.DNSContext,
+) (err error) {
 	if dctx.RequestedPrivateRDNS != (netip.Prefix{}) {
 		// Don't match client for private PTR request.
 		return p.Resolve(ctx, dctx)